feat(issue-2): move AstStatementCreator into Parser and move ctxToAst methods into inner class
